Fix support for flash with string keys (options)
  In `7b47a77' we added support for flash with "string keys", but we
focused on "top level" hash keys. `actionpack' 4.1 also changed behavior
for the *value* of a specific flash entry, the one with the key
"options". Even when we set some options explicitly with a symbol as the
key for a flash message, the helper in the view layer will transform the
symbol to a string.